1.Создание функции получения волонтёров с пагинацией с Dapper
Для Dapper: создание конфигурации для JsonB-столбцов, чтобы сразу в запросе маппить в объекты.

Создание вспомогательных таблиц через Tools Scaffold для удобства написания запросов.

2.Создание абстракции для команд, запросов и хэндлеров в Application

Добавление библиотеки Scrutor, рефакторинг Inject.

3.Рефакторинг Error

Добавление в Error свойства ValidationErrors[].

Замена List<Error> на просто Error в Result.

4.Добавление кастомного атрибута [Unique] в DomainModels

Добавление метода-расширения для EntityTypeBuilder

ApplyUniqueConstraints<>() для автоматической конфигурации правила уникальности.

(Только для значимых типов, string и ValueObjects, которые настроены через OwnsOne).

Рефакторинг проверки уникальности в хэндлерах

+++ src/PetFamily.Application/Commands/VolunteerManagment/DeleteVolunteer/DeleteVolunteerHandler.cs
@@ -1,7 +1,9 @@
 ﻿using Microsoft.Extensions.Logging;
 using Microsoft.Extensions.Options;
+using PetFamily.Application.Abstractions;
 using PetFamily.Application.Commands.FilesManagment;
 using PetFamily.Application.Commands.FilesManagment.Dtos;
+using PetFamily.Application.Commands.SharedCommands;
 using PetFamily.Application.IRepositories;
 using PetFamily.Domain.Results;
 
@@ -9,34 +11,36 @@ namespace PetFamily.Application.Commands.VolunteerManagment.DeleteVolunteer;
 
 public class DeleteVolunteerHandler(
     ILogger<DeleteVolunteerHandler> logger,
-    IVolunteerRepository volunteerRepository,
+    IVolunteerWriteRepository volunteerRepository,
     IFileRepository fileRepository,
     FilesProcessingQueue filesProcessingQueue,
-    IOptions<FileFolders> fileFoldersOptions)
+    IOptions<FileFolders> fileFoldersOptions) : ICommandHandler<Guid,VolunteerIdCommand>
 {
     private readonly IFileRepository _fileRepository = fileRepository;
-    private readonly IVolunteerRepository _volunteerRepository = volunteerRepository;
+    private readonly IVolunteerWriteRepository _volunteerRepository = volunteerRepository;
     private readonly ILogger<DeleteVolunteerHandler> _logger = logger;
     FilesProcessingQueue _filesProcessingQueue = filesProcessingQueue;
     FileFolders _fileFolders = fileFoldersOptions.Value;
-    public async Task<Result<Guid>> Handle(Guid volunteerId, CancellationToken cancelToken)
+    public async Task<Result<Guid>> Handle(VolunteerIdCommand command, CancellationToken cancelToken)
     {
-        var volunteer = await _volunteerRepository.GetByIdAsync(volunteerId, cancelToken);
-        var pets = volunteer.Pets;
+        var getVolunteer = await _volunteerRepository.GetByIdAsync(command.VolunteerId, cancelToken);
+        if (getVolunteer.IsFailure)
+            return Result.Fail(getVolunteer.Error);
+
+        var volunteer = getVolunteer.Data!;
+
         List<AppFile> imagesToDelete = [];
-        foreach (var pet in pets)
-        {
+
+        foreach (var pet in volunteer.Pets)
             imagesToDelete.AddRange(pet.Images.Select(x => new AppFile(x.Name,_fileFolders.Images)));   
-        }
+        
 
         await _volunteerRepository.Delete(volunteer, cancelToken);
 
         if(imagesToDelete.Count > 0)
-        {
-            await _filesProcessingQueue.DeleteChannel.Writer.WriteAsync(imagesToDelete,CancellationToken.None);
-        }
+            await _filesProcessingQueue.DeleteChannel.Writer.WriteAsync(imagesToDelete);       
 
-        _logger.LogInformation("Hard delete volunteer with id:{Id} successful!",volunteerId);
+        _logger.LogInformation("Hard delete volunteer with id:{Id} successful!", command.VolunteerId);
 
         return Result.Ok(volunteer.Id);
     }
